Backend/downloader/requirements.txt
anyio==3.6.2
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
fastapi==0.85.2
h11==0.14.0
httptools==0.5.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.24.1
packaging==22.0
psycopg2==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
pytube==12.1.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.20.4
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.19.0
watchfiles==0.18.0
websockets==10.4
